+++ ZeroClipboard.js
@@ -15,6 +15,7 @@
       return singleton;
     }
     this.handlers = {};
+    this._text = null;
     if (ZeroClipboard.detectFlashSupport()) this.bridge();
     if (query) this.glue(query);
     ZeroClipboard._client = this;
@@ -82,7 +83,7 @@
     delete ZeroClipboard.currentElement;
   };
   ZeroClipboard.Client.prototype.ready = function() {
-    return !!this.htmlBridge.getAttribute("data-clipboard-ready");
+    return this.htmlBridge.getAttribute("data-clipboard-ready") === "true";
   };
   function _getCursor(el) {
     var y = el.style.cursor;
@@ -100,9 +101,7 @@
   ZeroClipboard.Client.prototype.setCurrent = function(element) {
     ZeroClipboard.currentElement = element;
     this.reposition();
-    if (element.getAttribute("data-clipboard-text")) {
-      this.setText(element.getAttribute("data-clipboard-text"));
-    }
+    this.setText(this._text || element.getAttribute("data-clipboard-text"));
     if (element.getAttribute("title")) {
       this.setTitle(element.getAttribute("title"));
     }
@@ -124,9 +123,12 @@
   };
   ZeroClipboard.Client.prototype.setText = function(newText) {
     if (newText && newText !== "") {
-      this.htmlBridge.setAttribute("data-clipboard-text", newText);
+      this._text = newText;
       if (this.ready()) this.flashBridge.setText(newText);
     }
+  };
+  ZeroClipboard.Client.prototype.resetText = function() {
+    this._text = null;
   };
   ZeroClipboard.Client.prototype.setTitle = function(newTitle) {
     if (newTitle && newTitle !== "") this.htmlBridge.setAttribute("title", newTitle);
@@ -204,6 +206,9 @@
       break;
      case "mouseup":
       ZeroClipboard.currentElement.removeClass("zeroclipboard-is-active");
+      break;
+     case "complete":
+      this.resetText();
       break;
     }
     if (this.handlers[eventName]) {