+++ src/javascript/ZeroClipboard/client.js
@@ -16,6 +16,9 @@ ZeroClipboard.Client = function (query) {
   // event handlers
   this.handlers = {};
 
+  // text
+  this._text = null;
+
   // setup the flash->Javascript bridge
   if (ZeroClipboard.detectFlashSupport()) this.bridge();
 
@@ -168,7 +171,7 @@ ZeroClipboard.Client.prototype.resetBridge = function () {
  * returns true if the flash bridge is ready
  */
 ZeroClipboard.Client.prototype.ready = function () {
-  return !!this.htmlBridge.getAttribute("data-clipboard-ready");
+  return this.htmlBridge.getAttribute("data-clipboard-ready") === "true";
 };
 
 /*
@@ -209,10 +212,7 @@ ZeroClipboard.Client.prototype.setCurrent = function (element) {
 
   this.reposition();
 
-  // If the dom element contains data-clipboard-text set text
-  if (element.getAttribute("data-clipboard-text")) {
-    this.setText(element.getAttribute("data-clipboard-text"));
-  }
+  this.setText(this._text || element.getAttribute("data-clipboard-text"));
 
   // If the dom element has a title
   if (element.getAttribute("title")) {
@@ -256,9 +256,18 @@ ZeroClipboard.Client.prototype.reposition = function () {
  */
 ZeroClipboard.Client.prototype.setText = function (newText) {
   if (newText && newText !== "") {
-    this.htmlBridge.setAttribute("data-clipboard-text", newText);
+    this._text = newText;
     if (this.ready()) this.flashBridge.setText(newText);
   }
+};
+
+/*
+ * Sets the object's _text to null
+ *
+ * returns nothing
+ */
+ZeroClipboard.Client.prototype.resetText = function () {
+  this._text = null;
 };
 
 /*